feature complete

+++ Assets/Scripts/Enemies/Dummy/Dummy.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 namespace Enemies.Dummy {
     public class Dummy : Enemy {
-        public override void Hurt(int damage) {
+        public override void Hurt(float damage) {
             base.Hurt(damage);
             Debug.Log(this + " took damage " + damage);
         }
